+++ package.json
@@ -8,7 +8,7 @@
     "test": "jest --coverage",
     "format": "prettier --write ."
   },
-  "proxy": "http://localhost:4000",
+  "proxy": "http://localhost:3001",
   "husky": {
     "hooks": {
       "pre-commit": "npm run format"
