+++ src/components/Pages/Home/Home.jsx
@@ -1,55 +1,9 @@
 import React from 'react';
 import Title from '../../layout/Typography/Title/Title';
-import { useState, useEffect } from 'react';
-import axios from 'axios';
+import getLocalData from '../../../api/api';
 
 function Home() {
-  const [data, setdata] = useState('');
-  const [apiData, setApiData] = useState('');
-  const [currentLocation, setcurrentLocation] = useState();
-  const [weatherData, setweatherData] = useState();
-  const token =
-    'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9wZmEuZm9yZWNhLmNvbVwvYXV0aG9yaXplXC90b2tlbiIsImlhdCI6MTYzNzA1OTMyOCwiZXhwIjo5OTk5OTk5OTk5LCJuYmYiOjE2MzcwNTkzMjgsImp0aSI6IjYxNDdlMTEwY2Y2NzE1ZjAiLCJzdWIiOiJpdmFuc2hvc2hrb3YiLCJmbXQiOiJYRGNPaGpDNDArQUxqbFlUdGpiT2lBPT0ifQ.XZy1ejuB-5hwsFlyUvc0DoQw3DYN1iBZJptZidhoIzA';
-  useEffect(() => {
-    apiData
-      ? axios
-          .get(`https://pfa.foreca.com/api/v1/observation/latest/${apiData.data.id}`, {
-            headers: {
-              Authorization: `Bearer ${token}`
-            }
-          })
-          .then(result => setweatherData(result))
-      : null;
-  }, [apiData]);
-
-  useEffect(() => {
-    console.log('currentLocation', currentLocation);
-    currentLocation
-      ? axios
-          .get(
-            `https://pfa.foreca.com/api/v1/location/${currentLocation.long},${currentLocation.lat}`,
-            {
-              headers: {
-                Authorization: `Bearer ${token}`
-              }
-            }
-          )
-          .then(result => setApiData(result))
-      : null;
-  }, [currentLocation]);
-
-  useEffect(async () => {
-    navigator.geolocation.getCurrentPosition(pos => {
-      const newUserPos = {
-        lat: pos.coords.latitude,
-        long: pos.coords.longitude
-      };
-      setcurrentLocation(newUserPos);
-    });
-  }, []);
-
-  console.log('apiData FROM API: ', apiData);
-  console.log('weatherData', weatherData);
+  const weatherData = getLocalData();
 
   const WeatherReport = () => {
     if (!weatherData) return <p>Loading</p>;
